refactor: split engine-sync, index-based compositor loop, type canvas mock
- Split engine-sync.ts (680 lines) into three focused modules:
  sync-state.ts (per-engine TrackedState WeakMap), sync-layers.ts
  (the hot layer-diff path), engine-sync.ts (the remaining sync
  functions). Public API unchanged — consumers still import from
  engine-sync.

- Rewrite compositor::composite to iterate layer_stack by index
  instead of collecting a Vec<tuple> per frame. Only enabled effect
  descriptors are cloned; EffectsDesc is no longer cloned wholesale
  for layers without effects.

- Type canvas-mock's getContext override against the prototype's
  overloaded signature rather than casting through any.
